legal-workflow-be/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# BigQuery
google-cloud-bigquery==3.14.1

# Auth
PyJWT==2.10.1

# CORS
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Dev
python-dotenv==1.0.1
